Add an example integration with the material-components-web library
MDC Web helps developers to execute Material Design specifications.
This is a nice way to keep in sync with the latest MD specs without having
to reimplement the wheel.

MD is a proper choice as the foundation for Ketchup as it has been
developed at Google by a core team of UX designers and Engineers;
it brings a beautiful yet customizable design and works seamlessly
across devices.

MDC Web supports themes and several utilities exposed through JS
and SAAS mixins to make the customization process easy.

MDC for the Web is the successor to Material Design Lite.

Ref: https://github.com/material-components/material-components-web